requirements-legalize.txt
pyyaml>=6.0.1
google-adk>=0.1.0
litellm>=1.35.0
pydantic>=2.0.0
aiohttp>=3.10